Only attempt to `scrollTop` if element is available.
In later versions of Ember, the views and components can be stable
and are not guaranteed to be torn down before a new controller is
set on them.

In this case, the controller is initially set before the element has been
rendered to the DOM causing errors when invoking `closest` on undefined.