Includes Ghost-Editor 0.0.8
Refs #7429

Ghost Editor includes:
- Responsive toolbars
- Both Ember and Plain javascript cards
- An embeded HTML card

This is still an early release, but things are moving in the right direction. :)

We're still defining the spec for the UI, so expect drastic changes over the next couple of weeks.

This is going to be a great writing experience and we can't wait to show what we have planned.