refs #7429
Finally it's starting to feel like a real editor! Although there will be another version bump over the weekend which improves the toolbar behavior and usability, and enables image uploading.

- Added the start of a new toolbar, what we're (well I am) calling the Owesome bar, not to be confused with the Firefox Awesome bar. It's a cultural thing. (google "O for awesome" if you're curious 🎡 of 💰).
- The idea of dragging and dropping cards has been removed for now, although the code will still be in there as we will support dragging cards around within content fairly shortly. When apps are included a better card interface will be required for a large amount of app created content cards (Oh yeah!)
- Ghost Server now pulls in it's configuration from Ghost-Editor, this allows Ghost-Editor to:
A) Keep cards up to date, and
B) define what happens if a card is missing.
- The whole "cards in admin written in ember and cards in server written in javascript" thing is still very much a work in progress, it's kind of messy as we find the optimum solution (which isn't the current solution).

So yeah, this is a WIP not the final styling, not the final interactions, not the final anything... :)