🐛  Fixed datepicker bug for post dates in past
closes TryGhost/Ghost#8891

Having a very old post date, reverting it to a draft and then planning to publish/schedule it, resulted in the datepicker dropdown navigation to show a the wrong date. This was caused, because in the `gh-date-time-picker.js` component, we used only the `didReceiveAttrs()` hook to populate the needed `_date` property. Once, this property is passed to the `power-datepicker`'s navigation `dp.nav' it wouldn't get updated, when the value was changed.

In this PR, I added a property `_initalDate` to the `gh-date-time-picker` component, which will get populated in the `didReceiveAttrs()` hook with the same values as before my changes and used to render the date in the input field (which is read-only). I additionally added a `didUpdateAttrs()` hook, which then sets the `_date` property after any changes of the blog publish date and gets passed as the value for the `power-datepicker`. Any other changes on the publish date/time reflects this.

- adds `didUpdateAttrs()` to `gh-date-time-picker` component
- adds `_initialDate` property which is used in the read-only input fields and is populated in the `didReceiveAttrs()`