Addresses issue #9296 (unexpected publishing of post)
Also confirmed in ubuntu.

If a user is trying to use the split screen preview shortcut in
the admin client (CTRL+ALT+P for all platforms), this also calls
the publish shortcut. CmdOrCtrl will assign CTRL for the leading
key for any other platform other than Mac.

CTRL + ALT + P is an already assigned shortcut in all platforms at:
https://github.com/TryGhost/Ghost-Admin/blob/312bd68f4c30b8b919ea4c0911c98faaca55ff8f/app/components/gh-markdown-editor.js#L113

A better fix would probably be a centralised or user controllable
shortcut config.

But for now, this or only setting the shift alternative for platforms
other than mac should fix this.