Remove misleading leading slash on tag slug
I’m a little fuzzy on why it’s there at all, frankly. If it’s supposed
to model the tag’s URL, it’s failing because it lacks the leading
`/tag`. If it’s supposed to be part of supporting tag hierarchies (which
don’t look to be supported at this stage, though `tag.parent` exists?),
then a leading slash has a *little* merit, but is still not necessary—
`foo/bar` works about as well as `/foo/bar`. In all, I reckon it’s best
to remove the leading slash, at the very least until tag hierarchies are
supported.